started turret commands, added smartdashboard input for testing flywheel, deleted duplicate commands, set default commands for other subsystems
