+++ src/main/java/com/spartronics4915/frc2020/commands/LauncherCommands.java
@@ -2,6 +2,7 @@ package com.spartronics4915.frc2020.commands;
 
 import com.spartronics4915.frc2020.subsystems.Launcher;
 
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.FunctionalCommand;
 
@@ -28,7 +29,7 @@ public class LauncherCommands
         @Override
         public void initialize()
         {
-            mLauncher.setRPS(40);
+            mLauncher.setRPS(SmartDashboard.getNumber("Launcher/FlywheelRPS", 0));
         }
     
         // Called every time the scheduler runs while the command is scheduled.
@@ -52,6 +53,50 @@ public class LauncherCommands
             mLauncher.reset();
         }
     }
+
+    
+    public class TurretTest extends CommandBase
+    {
+    
+        Launcher mLauncher;
+    
+        // You should only use one subsystem per command. If multiple are needed, use a
+        // CommandGroup.
+        public TurretTest(Launcher launcher)
+        {
+            mLauncher = launcher;
+            addRequirements(launcher);
+        }
+    
+        // Called when the command is initially scheduled.
+        @Override
+        public void initialize()
+        {
+            mLauncher.setRPS(SmartDashboard.getNumber("Launcher/FlywheelRPS", 0));
+        }
+    
+        // Called every time the scheduler runs while the command is scheduled.
+        @Override
+        public void execute()
+        {
+            mLauncher.runFlywheel();
+        }
+    
+        // Returns true when the command should end.
+        @Override
+        public boolean isFinished()
+        {
+            return false;
+        }
+    
+        // Called once the command ends or is interrupted.
+        @Override
+        public void end(boolean interrupted)
+        {
+            mLauncher.reset();
+        }
+    }
+
 
     /*Default command of the launcher subsystem, makes the flywheel's target rps 0*/
     public class LauncherDefaultCommand extends CommandBase
